// public/angular/app/interfaces/INotifAviso.ts
export interface INotifAviso {
    aviso: number;
    notificacion: number;
    fecha_fija: string;    
    funcionario_fija: number;
    observaciones_fija: string;
    fecha_desfija: string;
    funcionario_desfija: number;
    observaciones_desfija: string;
    fecha_registra: string;
    usuario: number;
    usuario_desc: string;
}