// public/angular/app/interfaces/IResponsable.ts
export interface IResponsable {
    responsable: number;
    persona: number;
    tipo_doc: number;
    tipo_doc_desc: string;
    numero_doc: string;
    nombres_apellidos: string;
    cargo: number;
    cargo_desc: string;
    tipo_acto: number;
    tipo_acto_desc: string;
    desde: string;
    hasta: string;
    usuario: number;
    usuario_desc: string;
    fecha_registro: string;
}